Rearrange database functions and add comments to each
